Rename raredisease_params_file_creator add to cg-test.yaml
